Move utils out of commands folder
