+++ convert/commands/audio.py
@@ -1,7 +1,7 @@
 from .base import Base
 
-from .utils import let_user_pick
-from .utils import run_ffmpeg
+from ..converter_utils import let_user_pick
+from ..converter_utils import run_ffmpeg
 
 
 class Audio(Base):
